More facade services

+++ conference/conference-facade/conference-facade-impl/src/main/java/com/prodyna/academy/pac/conference/facade/ConferenceServiceImpl.java
@@ -13,6 +13,7 @@ import com.prodyna.academy.pac.conference.base.monitoring.interceptor.ServiceLog
 import com.prodyna.academy.pac.conference.conference.model.Conference;
 import com.prodyna.academy.pac.conference.conference.service.ConferenceCRUDService;
 import com.prodyna.academy.pac.conference.facade.service.ConferenceService;
+import com.prodyna.academy.pac.conference.speaker.model.Speaker;
 import com.prodyna.academy.pac.conference.talk.model.Talk;
 import com.prodyna.academy.pac.conference.talk.service.TalkCRUDService;
 
@@ -44,10 +45,13 @@ public class ConferenceServiceImpl implements ConferenceService {
 
 	/**
 	 * Validates the date fields of the conference.
+	 * 
 	 * @param conferenceF
-	 * @throws BusinessException if the date fields are invalid
+	 * @throws BusinessException
+	 *             if the date fields are invalid
 	 */
-	private void validateConference(Conference conference) throws BusinessException{
+	private void validateConference(Conference conference)
+			throws BusinessException {
 		if (conference.getEnd().before(conference.getStart())) {
 			throw new BusinessException(
 					"Conference start date is after its end date.");
@@ -56,7 +60,8 @@ public class ConferenceServiceImpl implements ConferenceService {
 		Interval interval = conference.buildInterval();
 		for (Talk talk : talks) {
 			Interval talkInterval = talk.buildInterval();
-			boolean talkIntervalInsideConferenceInterval = interval.contains(talkInterval);
+			boolean talkIntervalInsideConferenceInterval = interval
+					.contains(talkInterval);
 			if (!talkIntervalInsideConferenceInterval) {
 				throw new BusinessException(
 						"The conference has talks which are outside of the set conference duration: "
@@ -67,6 +72,20 @@ public class ConferenceServiceImpl implements ConferenceService {
 
 	@Override
 	public Conference deleteConference(int conferenceId) {
+		List<Talk> talksByConference = talkService
+				.getTalksByConference(conferenceId);
+		if (!talksByConference.isEmpty()) {
+			StringBuilder sb = new StringBuilder();
+			for (Talk talk : talksByConference) {
+				if (sb.length() > 0) {
+					sb.append(", ");
+				}
+				sb.append(talk.getName());
+			}
+			throw new BusinessException(
+					"Cannot delete the conference due to assigned talks: "
+							+ sb.toString());
+		}
 		return conferenceService.deleteConference(conferenceId);
 	}
